requirements.txt
# Use ONLY these for Vercel
onnxruntime
opencv-python-headless
numpy
pillow
fastapi
python-multipart